Added support for fetching data and Displaying and coping the type Data dict
