+++ WebApp/DS_Core/KeywordBasedSearch/SearchEngines.py
@@ -1,0 +1,11 @@
+# Surface Web Search Engines
+SW_DUCK_DUCK_GO_URL = 'https://duckduckgo.com/?q='
+SW_AHMIA_URL = 'https://ahmia.fi/search/?q='
+
+# Dark Web Search Engines
+DW_DUCK_DUCK_GO_URL = 'https://duckduckgogg42xjoc72x3sjasowoarfbgcmvfimaftt6twagswzczad.onion/?q='
+DW_AHMIA_URL = 'http://juhanurmihxlp77nkq76byazcldy2hlmovfu2epvl5ankdibsot4csyd.onion/search/?q='
+DW_BRAVE_URL = 'https://search.brave4u7jddbv7cyviptqjc7jusxh72uik7zt6adtckl5f4nwy2v72qd.onion/search?q='
+DW_DEEP_SEARCH_URL = 'http://search7tdrcvri22rieiwgi5g46qnwsesvnubqav2xakhezv4hjzkkad.onion/result.php?search='
+DW_TOR66_URL = 'http://tor66sewebgixwhcqfnp5inzp5x5uohhdy3kvtnyfxc2e5mxiuh34iid.onion/search?q='
+DW_TORGLE_URL = 'http://iy3544gmoeclh5de6gez2256v6pjh4omhpqdh2wpeeppjtvqmjhkfwad.onion/torgle/?query='
