+++ WebApp/Home/static/css/style.css
@@ -1,0 +1,183 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto');
+@import url('https://fonts.googleapis.com/css2?family=Fira+Code');
+@import url('https://fonts.googleapis.com/css2?family=Coda');
+
+* {
+	margin: 0%;
+	padding: 0%;
+	color: white;
+	font-family: 'Fira Code';
+}
+
+body {
+	background: radial-gradient(circle, rgb(174, 10, 10) 50%, rgba(39,39,39,1) 100%);
+}
+
+.navbar {
+	padding: 2%;
+	background-color: #171720;
+	align-items: center;
+	box-shadow: 0px 0px 14px 0px rgba(0,0,0,0.75);
+	flex-direction: column;
+}
+
+	.navbar ul {
+		display: flex;
+		justify-content: space-between;
+		list-style: none;
+		font-size: 18px;
+		width: 25%;
+		margin: 0%;
+		padding: 0%;
+		align-items: center;
+	}
+
+	.navbar h1 {
+		font-family: 'Coda';
+		font-size: 300%;
+		font-weight: bold;
+		margin: 1%;
+		text-transform: uppercase;
+		text-align: center;
+		letter-spacing: 5px;
+	}
+
+	.navbar li {
+		padding: 2%;
+		font-family: 'Roboto' !important;
+	}
+
+	.navbar a:hover {
+		text-decoration: none;
+		font-size: 20px;
+	}
+
+.input-group {
+	width: 75% !important;
+	margin: 1% auto 3% !important;
+}
+
+.btn-primary {
+	border-radius: 0px 5px 5px 0px !important;
+}
+
+h4 {
+	margin: 3% !important;
+	font-weight: bold !important;
+}
+
+.progress-wrapper {
+	margin: auto;
+	width: 75%;
+}
+
+#progress-bar {
+	border-radius: 10px;
+}
+
+.container {
+	margin-top: 4%;
+	width: 100%;
+}
+
+#progress-bar-message {
+	text-align: center;
+	width: 60%;
+	font-size: 1.25rem;
+	margin: auto;
+	margin-top: 2%;
+	color: #cb0e0e;
+	background-color: #cce5ff;
+	border-color: #b8daff;
+	border-radius: 0.25rem;
+}
+
+#celery-result {
+	margin: 2% auto;
+	text-align: center;
+	color: rgb(255 255 255 / 75%);
+}
+
+form > p {
+	font-weight: bold;
+}
+
+/* Card starts*/
+
+.card {
+	border-radius: 15px !important;
+	background-color: rgb(197, 105, 105) !important;
+	margin: 3% 7%;
+	box-shadow: 4px 4px 6px 0 rgba(0, 0, 0, 0.3), -2px -2px 3px 0 rgba(255, 255, 255, 0.3);
+	border: 3px solid var(--currentColor) !important;
+	backdrop-filter: blur( 4px );
+}
+
+.card-header {
+	background-color: rgb(151, 18, 18) !important;
+	padding: 1% 2% !important;
+	border-radius: 15px !important;
+	color: white;
+	font-size: 200%;
+	font-family: 'Roboto';
+	font-weight: bold;
+	box-shadow: 4px 4px 6px 0 rgba(0, 0, 0, 0.3);
+	text-shadow: 2px 2px 2px black;
+	text-transform: uppercase;
+}
+
+.card-body {
+	padding: 2% !important;
+	font-size: 120%;
+}
+
+/* Card ends*/
+
+.modal {
+	background-color: black;
+}
+
+#modal_text {
+	color: black;
+	font-weight: bold;
+}
+
+.JSON_Output_child{   
+	position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: -17px; /* Increase/Decrease this value for cross-browser compatibility */
+    overflow-y: scroll; 
+	max-height:500px;
+}
+
+.JSON_Output_child::-webkit-scrollbar { 
+    display: none;  /* Safari and Chrome */
+	-ms-overflow-style: none;  /* Internet Explorer 10+ */
+    scrollbar-width: none;  /* Firefox */
+}
+
+.JSON_Output_parent{
+	position:relative;
+	width: 100%;
+    height: 100%;
+	overflow: hidden;
+}
+
+#Copy_JSON_Btn{
+	opacity: 0.8;
+	position:absolute; 
+	top:120px;
+	right:150px;
+	z-index:1;
+	transition: 500ms linear;
+}
+
+#Copy_JSON_Btn:hover{
+	opacity: 1.0;
+	position:absolute; 
+	top:120px;
+	right:150px;
+	z-index:1;
+}
