--- WebApp/Home/templates/Home/base.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>{% block title%}Diggy Spidy {% endblock %}</title>
-</head>
-<body>
-    <h1>Diggy Spidy</h1>
-    <div id="content" ,name="content">
-        {% block content%}
-        This is content !
-        {% endblock %}
-    </div>
-</body>
-</html>
