--- WebApp/Home/templates/Home/home.html
@@ -1,7 +1,0 @@
-{% extends 'Home/base.html' %}
-
-{% block title %} Home {% endblock %}
-
-{% block content %}
-<h1> This is home page ! </h1>
-{% endblock %}
