+++ WebApp/Home/templates/dashboard.html
@@ -1,0 +1,7 @@
+{% extends 'base.html' %}
+
+{% block title %} Dashboard {% endblock %}
+
+{% block content %}
+<h1> This is Dashboard ! </h1>
+{% endblock %}
