+++ WebApp/Home/templates/index.html
@@ -1,0 +1,19 @@
+{% extends 'base.html' %}
+
+{% block title %}
+Home Page
+{% endblock %}
+
+{% block links %}
+{% endblock %}
+
+{% block content %}
+<div class="card">
+	<h2 class="card-header">Title</h2>
+
+	<div class="card-body">
+		<img src="{% static 'assets/logo.png'%}"/>
+	</div>
+</div>
+
+{% endblock %}
