+++ WebApp/Home/urls.py
@@ -2,5 +2,7 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("",views.index,name="index"),
+    path("",views.home,name="home"),
+    path("home/",views.home,name="home"),
+    path("dashboard/",views.dashboard,name="dashboard"),
 ]
